+++ src/main/java/de/dm/intellij/liferay/language/clientextension/ClientExtensionYamlSchemaFileProviderFactory.java
@@ -14,14 +14,18 @@ public class ClientExtensionYamlSchemaFileProviderFactory {
 	private static final String CUSTOM_ELEMENT = "customElement";
 	private static final String EDITOR_CONFIG_CONTRIBUTOR = "editorConfigContributor";
 	private static final String FDS_CELL_RENDERER = "fdsCellRenderer";
+	private static final String FDS_FILTER = "fdsFilter";
 	private static final String GLOBAL_CSS = "globalCSS";
 	private static final String GLOBAL_JS = "globalJS";
 	private static final String IFRAME = "iframe";
+	private static final String INSTANCE_SETTINGS = "instanceSettings";
 	private static final String JS_IMPORT_MAPS_ENTRY = "jsImportMapsEntry";
 	private static final String NOTIFICATION_TYPE = "notificationType";
 	private static final String OAUTH_APPLICATION_HEADLESS_SERVER = "oAuthApplicationHeadlessServer";
 	private static final String OAUTH_APPLICATION_USER_AGENT = "oAuthApplicationUserAgent";
 	private static final String OBJECT_ACTION = "objectAction";
+	private static final String OBJECT_VALIDATION_RULE = "objectValidationRule";
+	private static final String SITE_INITIALIZER = "siteInitializer";
 	private static final String STATIC_CONTENT = "staticContent";
 	private static final String THEME_CSS = "themeCSS";
 	private static final String THEME_FAVICON = "themeFavicon";
@@ -30,9 +34,9 @@ public class ClientExtensionYamlSchemaFileProviderFactory {
 	private static final String WORKFLOW_ACTION = "workflowAction";
 
 	private static final String[] KNOWN_CLIENT_EXTENSION_TYPES = new String[] {
-			BATCH, CUSTOM_ELEMENT, FDS_CELL_RENDERER, GLOBAL_CSS, GLOBAL_JS,
-			IFRAME, JS_IMPORT_MAPS_ENTRY, NOTIFICATION_TYPE, OAUTH_APPLICATION_HEADLESS_SERVER,
-			OAUTH_APPLICATION_USER_AGENT, OBJECT_ACTION, STATIC_CONTENT, THEME_CSS,
+			BATCH, CUSTOM_ELEMENT, FDS_CELL_RENDERER, FDS_FILTER, GLOBAL_CSS, GLOBAL_JS,
+			IFRAME, INSTANCE_SETTINGS, JS_IMPORT_MAPS_ENTRY, NOTIFICATION_TYPE, OAUTH_APPLICATION_HEADLESS_SERVER,
+			OAUTH_APPLICATION_USER_AGENT, OBJECT_ACTION, OBJECT_VALIDATION_RULE, STATIC_CONTENT, SITE_INITIALIZER, THEME_CSS,
 			THEME_FAVICON, THEME_JS, THEME_SPRITEMAP, WORKFLOW_ACTION
 	};
 
@@ -51,6 +55,9 @@ public class ClientExtensionYamlSchemaFileProviderFactory {
 						"Frontend data set cell renderer Client Extension", project, "client-extension-schema-fds-cell-renderer.json", type -> StringUtil.equals(type, FDS_CELL_RENDERER)
 				),
 				new GenericClientExtensionYamlSchemaFileProvider(
+						"Frontend Data Set Filter Client Extension", project, "client-extension-schema-fds-filter.json", type -> StringUtil.equals(type, FDS_FILTER)
+				),
+				new GenericClientExtensionYamlSchemaFileProvider(
 						"CSS Client Extension", project, "client-extension-schema-global-css.json", type -> StringUtil.equals(type, GLOBAL_CSS)
 				),
 				new GenericClientExtensionYamlSchemaFileProvider(
@@ -58,6 +65,9 @@ public class ClientExtensionYamlSchemaFileProviderFactory {
 				),
 				new GenericClientExtensionYamlSchemaFileProvider(
 						"IFrame Client Extension", project, "client-extension-schema-iframe.json", type -> StringUtil.equals(type, IFRAME)
+				),
+				new GenericClientExtensionYamlSchemaFileProvider(
+						"Instance Settings Client Extension", project, "client-extension-schema-instance-settings.json", type -> StringUtil.equals(type, INSTANCE_SETTINGS)
 				),
 				new GenericClientExtensionYamlSchemaFileProvider(
 						"JS Import Maps Entry Client Extension", project, "client-extension-schema-js-import-maps-entry.json", type -> StringUtil.equals(type, JS_IMPORT_MAPS_ENTRY)
@@ -73,6 +83,12 @@ public class ClientExtensionYamlSchemaFileProviderFactory {
 				),
 				new GenericClientExtensionYamlSchemaFileProvider(
 						"Object Action Client Extension", project, "client-extension-schema-object-action.json", type -> StringUtil.equals(type, OBJECT_ACTION)
+				),
+				new GenericClientExtensionYamlSchemaFileProvider(
+						"Object Validation Rule Client Extension", project, "client-extension-schema-object-validation-rule.json", type -> StringUtil.equals(type, OBJECT_VALIDATION_RULE)
+				),
+				new GenericClientExtensionYamlSchemaFileProvider(
+						"Site Initializer Client Extension", project, "client-extension-schema-site-initializer.json", type -> StringUtil.equals(type, SITE_INITIALIZER)
 				),
 				new GenericClientExtensionYamlSchemaFileProvider(
 						"Static content Client Extension", project, "client-extension-schema-static-content.json", type -> StringUtil.equals(type, STATIC_CONTENT)